% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{catholic_dioceses}
\alias{catholic_dioceses}
\title{Roman Catholic dioceses in the United States, Canada, and Mexico}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	425 obs. of  6 variables:
 $ diocese: chr  "Baltimore, Maryland" "New Orleans, Louisiana" "Boston, Massachusetts" "Louisville, Kentucky" ...
 $ rite   : chr  "Latin" "Latin" "Latin" "Latin" ...
 $ lat    : num  39.3 30 42.4 38.3 40.7 ...
 $ long   : num  -76.6 -90.1 -71.1 -85.8 -74 ...
 $ event  : Factor w/ 2 levels "erected","metropolitan": 1 1 1 1 1 1 1 1 1 1 ...
 $ date   : POSIXct, format: "1789-04-06" "1793-04-25" ...
}}
\usage{
catholic_dioceses
}
\description{
Dates when Roman Catholic dioceses and archdioceses in the United States were
founded or made metropolitan sees, with geocoded locations.
}
\section{Variables}{


  \itemize{

  \item \code{diocese}: name, and thus location, of the diocese.

  \item \code{event}, \code{date}: the date when the diocese was
  \code{erected}, that is first founded, or made a \code{metropolitan} see.
  Encoded as a date object.

  \item \code{rite}: the rite overseen by the diocese. Regions with ordinary
  jurisdiction but not episcopal character are not included.

  \item \code{lat}, \code{long}: latitude and longitude coordinates for the
  headquarters city of the diocese.

  }
}
\examples{
head(catholic_dioceses)
}
\references{
This data is compiled from several sources:

  Joseph Bernard Code, \emph{Dictionary of the American Hierarchy
  (1789-1964)} (New York: Joseph F. Wagner, 1964), 425-26.

  For the United States since 1963, Canada, and Mexicode:
  \url{http://www.catholic-hierarchy.org/}
}
\keyword{datasets}

