% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{historydata}
\alias{historydata}
\alias{historydata-package}
\title{Datasets for historians}
\description{
This package provides sample datasets of interest to historians, analogous to
the \code{datasets} package for R generally and the \code{histdata} package
for datasets from the history of statistics.
}
\section{References}{
 Citations to the sources for datasets are provided in
  the documentation for each dataset.
}

