% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodists.R
\docType{data}
\name{methodists}
\alias{methodists}
\title{Methodist Minutes of the Annual Conferences, 1786--1834}
\format{A data frame with 20241 observations of 11 variables.}
\usage{
methodists
}
\description{
This dataset contains transcriptions of the membership figures from the
annual meeting minutes of the Methodist Episcopal Church in the United States
of America.
}
\author{
\href{http://lincolnmullen.com}{Lincoln Mullen}, George Mason
  University
}
\references{
See the \code{url} field in each row for a link to the page in
  the Hathi Trust from which that row was transcribed.
}
\keyword{datasets}

