% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perseus_dictionary.R
\docType{data}
\name{perseus_dictionary}
\alias{perseus_dictionary}
\title{A Dictionary of Greek and Roman biography and mythology. William Smith, LLD, ed. 1890}
\format{A data frame with 2744 observations of two variables.}
\usage{
perseus_dictionary
}
\description{
This dataset contains entries on biographical and mythological figures from the Classical Era.
}
\section{Variables}{


  \itemize{

  \item \code{entry}: Text entry including name, description, and textual references.

  }
}

\examples{
head(perseus_dictionary)
}
\keyword{datasets}
