% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi-war.R
\docType{data}
\name{quasi_war}
\alias{quasi_war}
\title{Naval encounters during the Quasi War between France and the United States of
America}
\format{
A data frame with 198 observations of 16 variables.
}
\usage{
quasi_war
}
\description{
Naval encounters during the Quasi War between France and the United States of
America
}
\examples{
head(quasi_war)
}
\references{
This data was gathered by \href{http://abbymullen.org/}{Abby
  Mullen}.
}
\keyword{datasets}
