% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{tudors}
\alias{tudors}
\title{Tudor dynasty}
\format{A data frame with 35 observations of 3 variables.}
\usage{
tudors
}
\description{
Spousal and parent/child relationships among selected members of the Tudor
dynasty, suitable for network analysis.
}
\section{Variables}{


  \itemize{

  \item \code{person_1}, \code{person_2}: The two people in the relationship.

  \item \code{relationship}: The type of relationship. }
}
\examples{
data(tudors)
}
\keyword{datasets}

