% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_cities_pop.R
\docType{data}
\name{us_cities_pop}
\alias{us_cities_pop}
\title{United States Historical City Populations, 1790-2010}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 62224 rows and 17 columns.}
\usage{
us_cities_pop
}
\description{
A dataset of cities mentioned in the U.S. Census and other data sources. In
general, cities are included once they reach a population of 2,500. Data is
recorded every ten years. See
\href{https://github.com/cestastanford/historical-us-city-populations/}{the
dataset's GitHub page} for details on how the data was gathered.
}
\references{
Erik Steiner and Jason Heppler, "United States Historical City
  Populations, 1790-2010," Spatial History Project, Center for Spatial and
  Textual Analysis, Stanford University,
  \url{https://github.com/cestastanford/historical-us-city-populations/}
}
\keyword{datasets}

