% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us-national-population.R
\docType{data}
\name{us_national_population}
\alias{us_national_population}
\title{Population of the United States, 1790-2020}
\format{
A data frame with 24 observations of 2 variables.
}
\usage{
us_national_population
}
\description{
Population figures for the entire United States of America from the decennial census.
}
\section{Variables}{


  \itemize{

  \item \code{year}: date of the census.

  \item \code{population}: population of the state or territory.

  }
}

\examples{
head(us_national_population)
if(require(ggplot2)) {
  ggplot(us_national_population,
         aes(x = year, y = population)) +
  geom_line() +
  ggtitle("Population of the United States, 1790-2010")
}
}
\references{
This dataset has been gathered by the
  \href{https://www.nhgis.org/}{NHGIS}.
  Steven Manson, Jonathan Schroeder, David Van Riper, Katherine Knowles, Tracy Kugler,
  Finn Roberts, and Steven Ruggles.
  \emph{IPUMS National Historical Geographic Information System: Version 19.0}
  [dataset]. Minneapolis, MN: IPUMS, 2024.
  \doi{10.18128/D050.V19.0}
}
\keyword{datasets}
