% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/historydata-package.r
\docType{package}
\name{historydata}
\alias{historydata}
\alias{historydata-package}
\title{Datasets for historians}
\description{
This package provides sample datasets of interest to historians, analogous to
the \pkg{datasets} package for R generally and the \pkg{histdata} package for
datasets from the history of statistics.
}
\section{References}{
 Citations to the sources for datasets are provided in
  the documentation for each dataset.
}

