% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{sarna}
\alias{sarna}
\title{Population estimates of American Jews}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	23 obs. of  5 variables:
 $ year               : int  1660 1700 1776 1790 1800 1820 1830 1840 1850 1860 ...
 $ estimate_low       : int  50 200 1000 1300 2500 2650 4000 15000 50000 125000 ...
 $ estimate_high      : int  50 300 2500 3000 2500 3000 6000 15000 50000 200000 ...
 $ percentage_pop_low : num  NA NA 0.04 0.03 0.04 0.03 0.03 0.09 0.22 0.4 ...
 $ percentage_pop_high: num  NA NA 0.1 0.08 0.04 0.03 0.05 0.09 0.22 0.63 ...
}}
\usage{
sarna
}
\description{
This dataset contains estimates of the population of American Jews.
}
\section{Variables}{


  \itemize{

  \item \code{year}: date of estimate.

  \item \code{estimate_low}, \code{estimate_high}: lower and upper bounds of
  the population estimate in absolute terms.

  \item \code{percentage_pop_low}, \code{percentage_pop_high}: lower and
  upper bounds of percentage of American Jews in the total population of the
  United States.

  }
}

\section{Citation}{
 This data is taken from the appendix in Jonathan D. Sarna,
  \emph{American Judaism: A History} (New Haven: Yale University Press,
  2004), 375-376.
}
\examples{
head(sarna)
}
\keyword{datasets}

