% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/us-state-populations.R
\docType{data}
\name{us_state_populations}
\alias{us_state_populations}
\title{Populations of US states and territories, 1790-2010}
\format{A data frame with 983 observations of 4 variables.}
\usage{
us_state_populations
}
\description{
Population figures for US states and territories from the decennial census.
}
\section{Variables}{


  \itemize{

  \item \code{year}: date of the census.

  \item \code{state}: name of the state or territory.

  \item \code{population}: population of the state or territory.

  \item \code{GISJOIN}: a unique identifier for joining NHGIS data to spatial
  data.

  }
}
\examples{
head(us_state_populations)
}
\references{
This dataset has been gathered by the
  \href{https://www.nhgis.org/}{NHGIS}. Minnesota Population Center,
  \emph{National Historical Geographic Information System: Version 2.0}
  (Minneapolis: University of Minnesota, 2011).
}
\keyword{datasets}

